add three-team tiebreaking
when we check that exactly three teams have the best records
we try a new three-team breaker that enumerates all the
possible dyads and then checks the most important tie-breaks
first.

The search is "shallow" meaning we want to see head-to-head
results across all dyads before diving down to more esoteric
breakers in one dyad.

Any tiebreak is considered success and we have broken the 3-way
because the final two will still have to pass through the 2-way
tiebreak.

Thus it is possible that randomly one team will get pushed down
to third place because of the order of tiebreaks considered but
the ultimate winner will still be deserving.