cleaner way to calculate opponent margin test
Factor out the "opponents by margin" check since we need to do it
for both teams.

With that simplified we can just do it twice and then check for
intersections with a one-line counter.

This solves the problem of supporting multiple games with an
opponent in the new torture tests
schedule_maker.three_team_tie()
schedule_maker.two_team_tie_one_doormat()

Also passes the real-world checks and it's much cleaner.