track "no winner" in the simulation scoreboard
Having no resolvable winner is possible and we were ignoring that
chance in the scoreboard countups.

So now we do our full resolution in every simulation run and
use the result boolean to tell us if it should be "no winner"
which is tracked with a dedicated method in the scoreboard.

No winner gets a line item on the final printout.