housekeeping on tie scenarios
Fix a log bug: the log statement for interpreting dyads in the
three team tie was using the names from _after_ the order
reassignment. So now we pull out the teams being tested so we
can accurately log who was affected

For consistency, log when a tie break method succeeds in all
cases

Print the standings no matter what if we're verbose, if we
fail on tie breaks we still want to see the tied standings.